requirements.txt
apns2==0.4.1
asn1crypto==0.24.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.2.2
defusedxml==0.5.0
Django==2.0.7
django-allauth==0.36.0
django-appconf==1.0.1
django-auditlog==0.4.5
django-ckeditor==5.5.0
django-classy-tags==0.8.0
django-cors-headers==2.3.0
django-embed-video==1.1.2
django-filter==1.1.0
django-image-cropping==1.1.0
django-js-asset==1.1.0
django-jsonfield==1.0.1
django-push-notifications==1.6.0
django-rest-auth==0.9.3
django-taggit==0.23.0
django-taggit-templatetags2==1.6.1
djangorestframework==3.8.2
drf-base64==0.9.4
easy-thumbnails==2.5
h2==2.6.2
hpack==3.0.0
http-ece==1.0.5
hyper==0.7.0
hyperframe==3.2.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
oauthlib==2.1.0
Pillow==5.2.0
py-vapid==1.4.0
pycparser==2.18
PyJWT==1.6.4
python-dateutil==2.6.0
python3-openid==3.1.0
pytz==2018.5
pywebpush==1.7.0
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
xlrd==1.1.0
